pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "versioningit~=2.0"]
build-backend = "setuptools.build_meta"

# Self-descriptive entries which should always be present
# https://packaging.python.org/en/latest/specifications/declaring-project-metadata/
[project]
name = "QCANT"
description = "Quantum chemistry algorithms and acceleration utilities for near-term quantum computing research"
dynamic = ["version"]
readme = "README.md"
authors = [
    { name = "SRIVATHSAN P SUNDAR", email = "sundarbalaji4@gmail.com" }
]
license = "BSD-3-Clause"
license-files = ["LICENSE"]
# See https://pypi.org/classifiers/
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Chemistry",
    "Typing :: Typed",
]
requires-python = ">=3.11"
# Declare any run-time dependencies that should be installed with the package.
dependencies = [
    "numpy<2",
    "scipy<2",
    "pennylane",
    "pyscf",
    "openfermion",
    "autoray>=0.6.11,<0.9",
]


[project.urls]
Homepage = "https://github.com/asthanalab/QCANT"
Source = "https://github.com/asthanalab/QCANT"
Documentation = "https://asthanalab.github.io/QCANT/"

[project.optional-dependencies]
qulacs = [
  "qulacs",
]
gpu = [
  "pennylane-lightning-gpu",
  "cupy-cuda12x<14",
]
test = [
  "pytest>=6.1.2",
]
docs = [
    "sphinx",
    "pydata-sphinx-theme",
    "sphinx-design",
    "sphinx-copybutton",
]
dev = [
    "build",
    "twine",
    "packaging>=25",
    "pytest>=6.1.2",
    "pytest-cov",
    "sphinx",
    "pydata-sphinx-theme",
    "sphinx-design",
    "sphinx-copybutton",
    "matplotlib",
]

[tool.setuptools]
# This subkey is a beta stage development and keys may change in the future, see https://setuptools.pypa.io/en/latest/userguide/pyproject_config.html for more details
#
# As of version 0.971, mypy does not support type checking of installed zipped
# packages (because it does not actually import the Python packages).
# We declare the package not-zip-safe so that our type hints are also available
# when checking client code that uses our (installed) package.
# Ref:
# https://mypy.readthedocs.io/en/stable/installed_packages.html?highlight=zip#using-installed-packages-with-mypy-pep-561
zip-safe = false
# Let setuptools discover the package in the current directory,
# but be explicit about non-Python files.
# See also:
#   https://setuptools.pypa.io/en/latest/userguide/pyproject_config.html#setuptools-specific-configuration
# Note that behavior is currently evolving with respect to how to interpret the
# "data" and "tests" subdirectories. As of setuptools 63, both are automatically
# included if namespaces is true (default), even if the package is named explicitly
# (instead of using 'find'). With 'find', the 'tests' subpackage is discovered
# recursively because of its __init__.py file, but the data subdirectory is excluded
# with include-package-data = false and namespaces = false.
include-package-data = false
[tool.setuptools.packages.find]
namespaces = false
where = ["."]

# Ref https://setuptools.pypa.io/en/latest/userguide/datafiles.html#package-data
[tool.setuptools.package-data]
QCANT = [
    "py.typed"
]

[tool.versioningit]
default-version = "1+unknown"

[tool.versioningit.format]
distance = "{base_version}+{distance}.{vcs}{rev}"
dirty = "{base_version}+{distance}.{vcs}{rev}.dirty"
distance-dirty = "{base_version}+{distance}.{vcs}{rev}.dirty"

[tool.versioningit.vcs]
# The method key:
method = "git"  # <- The method name
# Parameters to pass to the method:
match = ["*"]
default-tag = "1.0.0"

[tool.versioningit.write]
file = "QCANT/_version.py"
